Minor protocol bug fixes...

+++ pilot/modules/protocol/protocol.h
@@ -38,9 +38,9 @@ void protocol_send_telemetry(vector_t attitude, double motor[]);
 // helper functions
 
 double protocol_byte_to_radian(uint8_t x);
-uint8_t protocol_radian_to_byte(float x);
+uint8_t protocol_radian_to_byte(double x);
 double protocol_byte_to_percent(uint8_t x);
-uint8_t protocol_percent_to_byte(float x);
+uint8_t protocol_percent_to_byte(double x);
 void protocol_double_to_bytes(double value, uint8_t *buf);
 double protocol_bytes_to_double(uint8_t *buf);
